Allow migration to field types with no length
Some MySQL field-types are being specified without a data length (e.g. BLOBs). Currently it is not possible to use such a custom column type in conjunction with the migrator since it will always try to create/modify the field like this, which is a syntax error:

````
ALTER TABLE `TableName` ADD COLUMN `icon` LONGBLOB(512) NULL
````

This change will allow the author of a model configuration to add a field like this (note the dataLength: false):
````
    "icon": {
      "type": "string",
      "dataType": "LONGBLOB",
      "mysql": {
        "dataType": "LONGBLOB",
        "dataLength": false
      }
    }
````